[RadixTree] Add minimal locking to RadixTree
The RadixTree class is *not* thread-safe for general usage. There is a per
node mutex that is locked on access to it's own EdgeMap to make inserts
and lookups thread safe for a specific access pattern where only
one thread at a time can extend an edge. Deletion is not thread-safe.