Added TrackingRadixTreeExt class
TrackingRadixTreeExt: "External" version of TrackingRadixTree that requires
minimal locking for concurrent access. This is a modified version of
TrackingRadixTree that eliminates the map of tracking objects to nodes by
storing a pointer to the Node in the TrackingObject itself.
NB: This class is only thread-safe for insertions. Removals can only safely
be done by a single thread with exlcusive access.