Fixed SHA256 bug.  Now runs concrete with: ECDHE-ECDSA-AES128-GCM-SHA256
Boringssl calls SHA256_Update with a null data arguement and 0 len.  As update
is supposed to copy len bytes from data and include it in the hash state
maintained by ctx.  It returns directly on 0 len (note, we depend on this
check in the bssl code, so if they ever remove it we will have issues)