Add special function cliver_tls_master_secret
Along with the command line option -tls-master-secret-file, this enables cliver
to read the TLS master secret from a dedicated 48-byte binary file, instead of
reading it from the ktest file. Currently, the model for the OpenSSL function
tls1_generate_master_secret will first try this dedicated file, and if that
fails (or the command line option was not used), fall back to ktest.  This
enables backwards compatibility with the existing buildbot experiments without
recompiling.