[multipass] bugfix: crash when no new symbolics in round
The initial implementation of multipass starts the second pass at the beginning
of the round.  We then optimized it (commit 8dc482b) to begin the second pass
at the creation of the first symbolic variable in the round. When this does not
occur, we hit assert(multi_pass_clone_ != NULL) and the verifier aborts.

Fix: revert to old behavior when no new symbolic variables are created in a
round, i.e., start the 2nd pass at the beginning of the round.