[multipass] bugfix: multipass fails to terminate
When concretizations are possible, but for symbolic variables created in a
*previous* round, the set of bindings is non-empty and remains so even after
further passes.  Unfortunately, the termination condition for multipass was
that the set of bindings is empty, not that it reaches a fixed point. In this
case, we would never terminate.

The fix is to keep around a copy of the bindings from the previous pass.  If it
has not changed (i.e., we have learned nothing more), then terminate.