[LazyConstraint] test for expected erroneous accepts w/ prohib funcs
When running with prohibitive functions (but NOT lazy constraint generators)
there are cases in which our behavioral verifier is expected to erroneously
accept a network trace. In particular these are the cases when a prohibitive
function is skipped, and the multipass algorithm does not return in a
subsequent pass to execute the function concretely.  This can happen, for
example, if the prohibitive function's inputs are not learned until a later
round.