Added ktest_dup, pam_setcred, and ktest_recvmsg_fd.
ktest_recvmsg_fd is a stand in for recvmsg, when recvmsg is recieving a
file descriptor.  This is not a very general model.

ktest_dup simply creates a new socket and adds it to the list of sockets
we're tracking for ktest_select.  So it doesn't have any relationship with
the fd it is supposed to be duplicating.  It also does nothing about returing
the lowest numbered fd possible.