We're calling CVExecutor::branch because we're encountering a select
Instruction. We do not encounter select instructions in the sparklee
run, and we do not call CVExecutor::branch.

I had to comment out an assert that there be only 1 condition.
Robby had a comment about cliver assuming that each branch only
produce 2 children.

It would be useful to eliminate these instructions, because then
we would be able to avoid propogating chopped's changes to
CVExecutor::branch.